requirements.txt
google-adk>=1.0.0
litellm>=1.65.5
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
tavily-python
